check for lt returned by GetCachedInstance being null
Under some very weird circunstances that I haven't been able to reproduce but I'm seeing the logs for, lt is null. 
Since GetCachedInstance returns a null in some instances I would also argue this check should be in place anyways to avoid the null reference exception when the conditions for that null return occur.