Fix for working with msgid with non-ASCII characters (e.g when using non-english language as base language)
Removed --omit-header, as, as it stated in docs it causes errors with non-ASCII msgids

Reference: http://www.gnu.org/software/gettext/manual/gettext.html#xgettext-Invocation
5.1.7 Output details:

‘--omit-header’
Don't write header with ‘msgid ""’ entry.
This is useful for testing purposes because it eliminates a source of variance for generated .gmo files. With --omit-header, two invocations of xgettext on the same files with the same options at different times are guaranteed to produce the same results.

Note that using this option will lead to an error if the resulting file would not entirely be in ASCII.